requirements.txt
numpy
plyfile
sklearn
scikit-image
torch
trimesh
PyOpenGL
